requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.9.5
django-ckeditor==5.4.0
django-debug-toolbar==1.9.1
django-extensions==2.0.5
django-js-asset==1.0.0
django-pure-pagination==0.3.0
ecdsa==0.13
facebook-sdk==2.0.0
idna==2.6
Markdown==2.6.11
Pillow==5.0.0
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
whitenoise==3.3.1
